adgm_corp_agent/requirements.txt
langchain>=0.0.340
chromadb>=0.3.28
sentence-transformers>=2.2.2
PyMuPDF>=1.22.5
python-docx>=0.8.11
streamlit>=1.25.0
ollama>=0.1.0   # optional: only if you use ollama python client; remove if not using
tqdm
